refactor(routes): renomear enum route-type para route.type e atualizar imports
