+++ nexus-backend/src/modules/routes/dto/create-route.dto.ts
@@ -18,7 +18,7 @@ import {
 import { ApiProperty, ApiPropertyOptional } from '@nestjs/swagger';
 import { Type } from 'class-transformer';
 import { RouteStatus } from '../enums/route-status';
-import { RouteType } from '../enums/route-type';
+import { RouteType } from '../enums/route.type';
 
 /**
  * DTO para criar parada de rota
