+++ nexus-backend/src/modules/routes/enums/route.type.ts
